+++ draft-js-side-toolbar-plugin/CHANGELOG.md
@@ -6,4 +6,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### 2.0.2
 - accomodate editor dom reference for both react 0.15 and 0.16
 
+- Fixed positioning when the editor has a parent node that is positioned (i.e. `relative`, `absolute` or `fixed`).
+
 ### Released the first working of DraftJS Side Toolbar Plugin